chore: go mod tidy
