--- pkg/did/fspath.go
@@ -1,1 +1,0 @@
-package did
