+++ pkg/vault/vault.go
@@ -3,6 +3,7 @@ package vault
 import (
 	"context"
 
+	"github.com/ipfs/boxo/ipns"
 	"github.com/ipfs/kubo/client/rpc"
 	iface "github.com/ipfs/kubo/core/coreiface"
 	"github.com/ipfs/kubo/core/coreiface/options"
@@ -26,6 +27,7 @@ type Vault struct {
 	Key         iface.Key
 	SonrAddress string
 	PeerID      peer.ID
+	IPNS        ipns.Name
 }
 
 func New(ctx context.Context) (*Vault, error) {
@@ -34,14 +36,22 @@ func New(ctx context.Context) (*Vault, error) {
 	if err != nil {
 		return nil, err
 	}
-	key, err := c.Key().Generate(ctx, kc.Address, options.Key.Type(options.Ed25519Key))
+	key, err := c.Key().Generate(context.Background(), kc.Address, options.Key.Type(options.Ed25519Key))
 	if err != nil {
 		return nil, err
 	}
-
+	path, err := c.Unixfs().Add(context.Background(), kc.Directory)
+	if err != nil {
+		return nil, err
+	}
+	name, err := c.Name().Publish(context.Background(), path, options.Name.Key(key.ID().String()))
+	if err != nil {
+		return nil, err
+	}
 	return &Vault{
 		Key:         key,
 		SonrAddress: kc.Address,
 		PeerID:      key.ID(),
+		IPNS:        name,
 	}, nil
 }
